Simplified GUI input processing
  Having two seperate methods to handle input before and
after passing it to child components was unnecessarily
complicated. And letting the base GUIComponent class
define the order in which input is passed to child
components provided little flexibility for components. It
just wasn't a good system.
  Components now simply implement processInputEvent()
themselves. This means that common functionality will be
duplicated across classes and that consistent
functionality will not be guaranteed. However this should
make GUI input processing simpler to code and allow
components more flexibility in how they process input.