GUI components store an ID with which they are referenced
* Components can be retrieved from their parents with
their ID. This means that direct references to components
don't have to be stored and passed all around the code.
Unfortunately, this means that many operations with
components won't be type-safe, and many errors won't be
picked up by the compiler. Also, it might require that
individual component IDs have to be kept track of and
looked up during development. GUI creation and
interaction should be simpler, though, hopefully.

* Component positions and dimensions no longer have to be
specified in the constructor, in case it's more
intuitive to specify them later.

* Buttons and Windows use TextureRegions instead of Textures