Moved GUI code out of MenuScene into MenuSceneGUI
  MenuScene was getting bloated with a ton of GUI code.
Mostly just GUI creation. I'm not sure if simply moving
all of it out into a "MenuSceneGUI" is the right idea,
but detailed GUI creation should definitely be pulled out
of the scene itself somehow. The scene should be focused
on loading its major elements and handling logic.
  Maybe scenes should use pre-built, specialized GUI
components, such as a "MusicControlWindow" or a
"GPSControlWindow" instead of putting those together itself?
Or maybe it should use pre-built GUIs that manage entire
sets of components similar to the "MenuSceneGUI" in use
right now?
  Either way, MenuSceneGUI works fine for the time being
and keeps the scene class clean and unencumbered with all
the lengthy GUI creation code.